requirements.txt
mysql
mysql-connector
boto3
python-dotenv
